pyproject.toml
# =============================================================================
# BUILD SYSTEM
# =============================================================================
[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

# =============================================================================
# PROJECT METADATA
# =============================================================================
[project]
name = "plex_pipe"
requires-python = ">=3.11"
dynamic = ["version"]
description = "A pipeline for analyzing multiplexed data."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Kasia Kedziora", email = "kasia.kedziora@pitt.edu" }
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent"
]

# =============================================================================
# DEPENDENCIES
# =============================================================================

dependencies = [
    "anndata>=0.11.4",
    "dask[array]>=2024.4",
    "imagecodecs>=2024.9.22",
    "loguru>=0.7.3",
    "numpy>2.0",
    "opencv-python>=4.11",
    "pandas>=2.3.0",
    "pydantic>=2.7.6",
    "pyyaml>=6.0.2",
    "scikit-image>=0.25.2",
    "scipy>=1.15.3",
    "shapely>=2.1.1",
    "spatialdata>0.7",
    "tifffile>=2025.1.10",
    "zarr>3.0",
    "fsspec>=2024.6.1",
]

[project.optional-dependencies]
jupyter = [
    "ipykernel>=6.0",
    "ipywidgets>=8.1.8",
]
globus = [
    "globus-sdk>= 3.59.0",
]
segmentation-cpu= [
    "torch",
    "torchvision",
    "scanpy",
    "cellpose>=4.0",
    "instanseg-torch>=0.0.9",
]
segmentation-gpu= [
    "torch",
    "torchvision",
    "cellpose>=4.0",
    "instanseg-torch>=0.0.9",
]
gui = [
    "qtpy>=2.4.0",
    "PyQt5>=5.15.2",
    "napari>=0.6.6",
    "napari-spatialdata>0.5.7",
]
all-cpu = [
    "plex_pipe[jupyter,globus,gui,segmentation-cpu]"
]
all-gpu = [
    "plex_pipe[jupyter,globus,gui,segmentation-gpu]"
]

[dependency-groups]
all-dev-gpu = [
    "black>=23.3.0",
    "ruff>=0.0.241",
    "tox>=4.12.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "pre-commit>=3.6.0",
    "plex_pipe[all-gpu]",
]

# =============================================================================
# UV CONFIGURATION
# =============================================================================

[tool.uv]
override-dependencies = [
    "PyQt5==5.15.10",
    "PyQt5-Qt5==5.15.2"
]

[[tool.uv.index]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-gpu", marker = "extra == 'segmentation-gpu'" },
]
torchvision = [
  { index = "pytorch-gpu", marker = "extra == 'segmentation-gpu'" },
]

# =============================================================================
# TOOL CONFIGURATION
# =============================================================================
[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml"]

[tool.setuptools_scm]
write_to = "src/plex_pipe/_version.py"

[tool.black]
line-length = 79
target-version = ['py311', 'py312', 'py313']

[tool.ruff]
line-length = 79
lint.select = [
    "E", "F", "W", #flake8
    "UP", # pyupgrade
    "I", # isort
    "BLE", # flake8-blind-exception
    "B", # flake8-bugbear
    "A", # flake8-builtins
    "C4", # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "G", # flake8-logging-format
    "PIE", # flake8-pie
    "SIM", # flake8-simplify
]
lint.ignore = [
    "E501", # line too long. let black handle this
    "UP006", "UP007", # type annotation.
    "SIM117", # flake8-simplify
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".mypy_cache",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "*vendored*",
    "*_vendor*",
    "*.ipynb",
]
